separated processing data from component +  proper table print
